test: move the integration-test to test folder
